requirements.txt
asgiref==3.2.9
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==39.0.0
cycler==0.11.0
Django==3.0.7
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-matplotlib==0.1
django-multiselectfield==0.1.12
django-quill-editor==0.1.19
django-request==1.5.5
django-sortedm2m==3.1.1
django-tinymce==3.4.0
djangorestframework==3.12.4
dnspython==1.16.0
docutils==0.16
email-validator==1.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
fonttools==4.28.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonfield==3.1.0
Kivy==1.11.1
kivy-deps.angle==0.1.10
kivy-deps.glew==0.1.12
kivy-deps.gstreamer==0.1.18
kivy-deps.sdl2==0.1.23
Kivy-examples==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.3.2
lxml==4.6.4
MarkupSafe==1.1.1
matplotlib==3.5.0
monotonic==1.6
msgpack==1.0.4
multitasking==0.0.10
mysql-connector-python==8.0.31
newsapi-python==0.2.6
numpy==1.21.4
oauthlib==3.2.2
packaging==21.3
pandas==1.3.4
pandas-datareader==0.9.0
passlib==1.7.4
Pillow==8.4.0
prompt-toolkit==3.0.36
protobuf==3.20.1
pyasn1==0.4.8
pycparser==2.21
pygame==1.9.6
Pygments==2.6.1
PyMySQL==1.0.2
pyparsing==3.0.6
pypiwin32==223
pyrsistent==0.19.3
python-dateutil==2.8.1
python-slugify==7.0.0
pytz==2022.7
pywin32==228
PyYAML==6.0
redis==3.5.3
requests==2.24.0
rfc3987==1.3.8
rsa==4.9
selenium==3.141.0
selenium-driver-updater==5.1.7
semantic-version==2.10.0
setuptools-scm==6.3.2
simplejson==3.17.0
six==1.15.0
sleekxmpp==1.3.2
soupsieve==2.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
strict-rfc3339==0.7
text-unidecode==1.3
tomli==1.2.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.25.9
vine==5.0.0
wcwidth==0.2.5
webcolors==1.12
webull==0.3.0
Werkzeug==1.0.1
wget==3.2
WTForms==2.3.1
xmltodict==0.13.0
yahoo-finance==1.4.0
yfinance==0.1.67
zipp==3.11.0
